Update simple-upnpd dir

+++ simple-upnpd/install.sh
@@ -6,7 +6,7 @@ SPATH=$(dirname $0)
 
 set -e
 
-if [ -f "$SPATH/simple-upnpd" ]; then
+if [ -f "$SPATH/simple-upnpd" ] && [ -f "$SPATH/simple-upnpd.service" ]; then
 	cp -v $SPATH/simple-upnpd /usr/local/bin/simple-upnpd && chmod 755 /usr/local/bin/simple-upnpd
 	
 	if [ -f "$SPATH/simple-upnpd.xml" ] && [ ! -f "/etc/simple-upnpd.xml" ]; then
@@ -16,8 +16,6 @@ if [ -f "$SPATH/simple-upnpd" ]; then
 		sed "s@<UDN>.*</UDN>@<UDN>uuid:$(cat /proc/sys/kernel/random/uuid)</UDN>@" -i /etc/simple-upnpd.xml
 		
 		command -v dos2unix >/dev/null 2>&1 && dos2unix /etc/simple-upnpd.xml
-	elif [ -f "$SPATH/description.xml" ] && [ ! -f "/etc/simple-upnpd.xml" ]; then
-		cp -v $SPATH/description.xml /etc/simple-upnpd.xml && chmod 644 /etc/simple-upnpd.xml
 	fi
 	
 	cp -v $SPATH/simple-upnpd.service /etc/systemd/system && chmod 644 /etc/systemd/system/simple-upnpd.service
